split app.config elements to multiple files
