+++ TechTalk.SpecFlow/PlatformSpecific/AppConfig/BindingCultureConfigElement.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Configuration;
+
+namespace TechTalk.SpecFlow.Configuration
+{
+    public class BindingCultureConfigElement : ConfigurationElement
+    {
+        [ConfigurationProperty("name", DefaultValue = "en-US", IsRequired = false)]
+        [RegexStringValidator(@"\w{2}(-\w{2})?")]
+        public string Name
+        {
+            get { return (String)this["name"]; }
+            set { this["name"] = value; }
+        }
+    }
+}
