+++ TechTalk.SpecFlow/PlatformSpecific/AppConfig/StepAssemblyCollection.cs
@@ -1,0 +1,17 @@
+﻿using System.Configuration;
+
+namespace TechTalk.SpecFlow.Configuration
+{
+    public class StepAssemblyCollection : ConfigurationElementCollection
+    {
+        protected override ConfigurationElement CreateNewElement()
+        {
+            return new StepAssemblyConfigElement();
+        }
+
+        protected override object GetElementKey(ConfigurationElement element)
+        {
+            return ((StepAssemblyConfigElement)element).Assembly;
+        }
+    }
+}
